added jsx files for react components
