+++ LB Dungen/client/src/components/Footer.jsx
@@ -1,0 +1,1 @@
+//Add footer info here to be rendered
