+++ LB Dungen/client/src/components/Navbar.jsx
@@ -1,0 +1,15 @@
+//classes are bootstrap (used in class example, can be changed)
+//map through array of links to different page views on the site vita nav links in nav component
+export default function Nav({ links }) {
+    return (
+      <nav className="navbar navbar-expand-lg bg-secondary">
+        <div className="container-fluid">
+          <div className="collapse navbar-collapse" id="navbarSupportedContent">
+            <ul className="navbar-nav me-auto mb-2 mb-lg-0">
+              {links.map((link) => link)}
+            </ul>
+          </div>
+        </div>
+      </nav>
+    );
+  }
