+++ LB Dungen/client/src/pages/ErrorPage.jsx
@@ -1,0 +1,1 @@
+//need error handling page display here (see class example)
